+++ src/components/Pokemons/index.js
@@ -1,0 +1,1 @@
+export * from './Pokemons';
